+++ src/Polhem.OAuth2/Auth0/Auth0OAuth2Provider.cs
@@ -11,8 +11,8 @@ namespace Polhem.OAuth2
         /// Initializes a new instance of the <see cref="Auth0OAuth2Provider"/> class.
         /// </summary>
         /// <param name="options">The Auth0 OAuth2 options.</param>
-        /// <param name="httpClient">The HTTP client for requests to the provider, or null to use a shared instance.</param>
-        public Auth0OAuth2Provider(Auth0OAuth2Options options, HttpClient? httpClient = null) : base(options, httpClient)
+        /// <param name="httpClientFactory">Returns the HTTP client for a request to the provider, or null to use a shared instance.</param>
+        public Auth0OAuth2Provider(Auth0OAuth2Options options, Func<HttpClient>? httpClientFactory = null) : base(options, httpClientFactory)
         {
         }