+++ src/Polhem.OAuth2/Facebook/FacebookOAuth2Provider.cs
@@ -13,8 +13,8 @@ namespace Polhem.OAuth2
         /// Initializes a new instance of the <see cref="FacebookOAuth2Provider"/> class.
         /// </summary>
         /// <param name="options">The Facebook OAuth2 options.</param>
-        /// <param name="httpClient">The HTTP client for requests to the provider, or null to use a shared instance.</param>
-        public FacebookOAuth2Provider(FacebookOAuth2Options options, HttpClient? httpClient = null) : base(options, httpClient)
+        /// <param name="httpClientFactory">Returns the HTTP client for a request to the provider, or null to use a shared instance.</param>
+        public FacebookOAuth2Provider(FacebookOAuth2Options options, Func<HttpClient>? httpClientFactory = null) : base(options, httpClientFactory)
         {
         }